fix [swe] : Moved queue structs to states lib dir
Was having issues with the PlatformIO LDF compiler not recognising the compilation units correctly. Most likely will have to consider moving these again to the src folder once finishing the rest of the library objects

+++ software/Simple Transport Midi Controller/lib/HAL/HAL_RotaryEncoder.h
@@ -35,7 +35,7 @@ struct RotaryEncoderEvent {
 
     PushButtonEvent btnEvent = {};
     uint8_t encValue = 0;
-    int8_t delta = 0;
+    int16_t delta = 0;
 
 };
